+++ pkg/src/attack1.rs
@@ -18,7 +18,7 @@ impl CollisionAttackAgainstAnRT {
         self.attack_times.push(d.clock.elapsed().as_nanos());
         self.success = true;
         for i in 0..self.nwords_inj {
-            let w = Word::new_data(i as u16);
+            let w = Word::new_data(i as u32);
             d.log(
                 WRD_EMPTY,
                 ErrMsg::MsgAttk(format!("Sent Fake Data {} ", w).to_string()),
